+++ server/apps/sites/serializers/__init__.py
@@ -2,8 +2,7 @@ from .address_ser import AddressSerializer
 from .contact_ser import ContactSerializer, EpaPhoneSerializer
 from .epa_profile_ser import (
     EpaPermissionSerializer,
-    EpaProfileGetSerializer,
-    EpaProfileUpdateSerializer,
+    EpaProfileSerializer,
     SitePermissionSerializer,
 )
 from .site_ser import EpaSiteSerializer, SiteSerializer
